feat(object-storage): delete bucket in dashboard
ref: #DATATR-2744
