+++ packages/manager/apps/pci-object-storage/public/translations/pci-object-storage/storages/s3/dashboard/Messages_en_GB.json
@@ -65,5 +65,6 @@
   "editObjectLockToastSuccessDescription": "Object Lock options have been updated",
   "errorDurationMin": "The minimum period is 1 day.",
   "errorDurationMax": "The maximum period is 100 years or 36,500 days.",
-  "errorRuleRequired": "A retention policy is required to enable the default Object Lock."
+  "errorRuleRequired": "A retention policy is required to enable the default Object Lock.",
+  "deleteBucketButton": "Delete the bucket"
 }
