+++ packages/manager/apps/pci-object-storage/public/translations/pci-object-storage/storages/s3/dashboard/Messages_it_IT.json
@@ -65,5 +65,6 @@
   "editObjectLockToastSuccessDescription": "Le opzioni di Object Lock sono state aggiornate",
   "errorDurationMin": "La durata minima è di 1 giorno.",
   "errorDurationMax": "La durata massima è di 100 anni o 36.500 giorni.",
-  "errorRuleRequired": "Per attivare Object Lock di default è necessario definire una regola di retention."
+  "errorRuleRequired": "Per attivare Object Lock di default è necessario definire una regola di retention.",
+  "deleteBucketButton": "Elimina il bucket"
 }
