+++ packages/manager/apps/pci-object-storage/src/pages/object-storage/storage/swiftId/dashboard/_components/Configuration.component.tsx
@@ -1,19 +1,11 @@
-import {
-  Button,
-  Clipboard,
-  Tooltip,
-  TooltipProvider,
-  TooltipTrigger,
-  useToast,
-} from '@datatr-ux/uxlib';
+import { Clipboard, useToast } from '@datatr-ux/uxlib';
 import { useTranslation } from 'react-i18next';
-import { useNavigate, useParams } from 'react-router-dom';
+import { useParams } from 'react-router-dom';
 
 const Configuration = () => {
   const { t } = useTranslation('pci-object-storage/storages/swift');
   const toast = useToast();
   const { swiftId } = useParams();
-  const navigate = useNavigate();
   const onCopy = () => toast.toast({ title: t('copyLabel') });
 
   return (
@@ -22,23 +14,6 @@ const Configuration = () => {
         <h5>{t('swiftIdLabel')}</h5>
         <Clipboard value={swiftId} onCopy={onCopy} />
       </div>
-      <TooltipProvider>
-        <Tooltip>
-          <TooltipTrigger asChild>
-            <span className="inline-block w-full" tabIndex={0}>
-              <Button
-                data-testid="swift-config-delete-button"
-                variant="critical"
-                mode="outline"
-                className="w-full mt-4"
-                onClick={() => navigate('./delete')}
-              >
-                {t('deleteSwiftButton')}
-              </Button>
-            </span>
-          </TooltipTrigger>
-        </Tooltip>
-      </TooltipProvider>
     </div>
   );
 };
