+++ packages/manager/apps/pci-object-storage/src/pages/object-storage/storage/swiftId/dashboard/_components/InformationsDetails.component.tsx
@@ -1,6 +1,7 @@
 import { useTranslation } from 'react-i18next';
+import { useNavigate } from 'react-router-dom';
 import { Region } from '@datatr-ux/ovhcloud-types/cloud';
-import { Clipboard, useToast } from '@datatr-ux/uxlib';
+import { Button, Clipboard, useToast } from '@datatr-ux/uxlib';
 import storages from '@/types/Storages';
 import { useLocaleBytesConverter } from '@/hooks/useLocaleByteConverter.hook';
 
@@ -18,6 +19,7 @@ const InformationsDetails = ({
   const localeBytesConverter = useLocaleBytesConverter();
   const { t } = useTranslation('pci-object-storage/storages/swift');
   const toast = useToast();
+  const navigate = useNavigate();
   const onCopy = () => toast.toast({ title: t('copyLabel') });
 
   return (
@@ -39,6 +41,15 @@ const InformationsDetails = ({
         <h5>{t('spaceUsedTableHeader')}</h5>
         <p>{localeBytesConverter(swift.storedBytes, true, 2)}</p>
       </div>
+      <Button
+        data-testid="swift-container-delete-button"
+        variant="critical"
+        mode="outline"
+        className="mt-4"
+        onClick={() => navigate('./delete')}
+      >
+        {t('deleteSwiftButton')}
+      </Button>
     </div>
   );
 };
